Split components of trivia game, it is still not functional.
