+++ door/commands/trivia/data.py
@@ -1,56 +1,10 @@
-"""
-questions from https://github.com/uberspot/OpenTriviaQA/
-"""
-
 from pathlib import Path
-import threading
 import sqlite3
 
-
-from pydantic import BaseModel
 from loguru import logger as log
+from pydantic import BaseModel
 
-from . import BaseCommand, CommandLoadError, CommandRunError
-
-
-DDL = [
-    """
-CREATE TABLE IF NOT EXISTS category (
-  id INTEGER PRIMARY KEY AUTOINCREMENT,
-  name TEXT
-);""",
-    """
-CREATE TABLE IF NOT EXISTS question (
-  id INTEGER PRIMARY KEY AUTOINCREMENT,
-  category INTEGER,
-  question TEXT,
-  FOREIGN KEY(category) REFERENCES category(id)
-);""",
-    """
-CREATE TABLE IF NOT EXISTS answer (
-  question INTEGER,
-  choice TEXT, -- a, b, c, etc
-  answer TEXT,
-  correct BOOLEAN,
-  FOREIGN KEY(question) REFERENCES question(id)
-)""",
-    """
-CREATE TABLE IF NOT EXISTS user (
-  userid TEXT PRIMARY KEY,
-  pending_question INTEGER NULL
-  FOREIGN KEY(pending_question) REFERENCES question(id)
-);""",
-    """
-CREATE TABLE IF NOT EXISTS response (
-  userid TEXT,
-  question INTEGER,
-  answer INTEGER,
-  timestamp INTEGER,
-  correct BOOLEAN,
-  FOREIGN KEY (userid) REFERENCES user(userid)
-);
-""",
-]
+from .. import CommandLoadError
 
 
 class Answer(BaseModel):
@@ -142,9 +96,11 @@ def create_database(db_path: Path, questions_path: Path, load_questions: bool = 
     create tables and populate with categories, questions, and answers
     """
     db = sqlite3.connect(db_path)
-    cursor = db.cursor()
-    for statement in DDL:
-        cursor.execute(statement)
+
+    # run SQL
+    ddl_file = Path(__file__).with_name("trivia.sql")
+    db.executescript(ddl_file.open("r").read())
+    db.commit()
 
     if load_questions is False:
         return
@@ -153,6 +109,8 @@ def create_database(db_path: Path, questions_path: Path, load_questions: bool = 
     category: Category
     question: Question
     answer: Answer
+
+    cursor = db.cursor()
 
     log.debug("Writing questions to database..")
     for category in categories:
@@ -224,122 +182,3 @@ GROUP BY categories.name
     db.close()
     log.debug(f"Category counts: {category_counts}")
 
-
-class TriviaGame:
-    def __init__(self, cursor: sqlite3.Cursor, node: str):
-        self.cursor = cursor
-        self.node = node
-
-    def run(self, msg: str) -> str:
-        """
-        parse command
-        decide if we should ask or answer
-        return response
-        """
-
-        if msg.strip() == "":
-            return self.ask()
-
-        # TODO this is dumb
-        elif len(msg) == 1:
-            return self.answer()
-
-        else:
-            return "Unknown command"
-
-    def ask(self):
-        """
-        pick a new question
-        store asked question in db
-        return question text
-        """
-
-        # fetch a new question
-        result = self.cursor.execute(
-            """
-SELECT id, category, question
-FROM questions WHERE id NOT IN (SELECT question FROM responses WHERE userid = ?)
-ORDER BY RANDOM() LIMIT 1
-""",
-            (self.node,),
-        )
-        question_id, category_id, question = result.fetchone()
-
-        # record that we are asking it with an UPSERT
-        result = self.cursor.execute(
-            """
-INSERT INTO user (userid, pending_question) VALUES (:node, :question_id)
-ON CONFLICT (userid)
-DO UPDATE SET pending_question = :question_id
-""",
-            {"node": self.node, "question_id": question_id},
-        )
-
-        # 
-
-    def answer(self, user: str, answer: str):
-        "check letter answer 'a', 'b', 'c', 'd',"
-
-        # if there is no pending question
-        # if answer is correct or not
-        # record result
-        # return result and correct answer if wrong
-
-
-class TriviaCommand(BaseCommand):
-    """
-    idea:
-    'trivia' asks a new question (and skips the last)
-    'trivia last' restates the last question (if present), or returns a new one
-    'trivia stats' shows count of correct/incorrect
-    'trivia clear' clears all responses
-    'trivia categories' shows a list of categories
-
-    to start, just ask and get response
-    """
-
-    command = "trivia"
-    description = "answer trivia questions"
-    help = "Commands: 'skip', 'clear', 'last', 'stats'.\n\nAnswer questions with 'a', 'b', 'c', or 'd'."
-
-    def __init__(self):
-        # acquire lock before writing to the database
-        self.lock = threading.Lock()
-
-    def load(self):
-        self.db_file = Path(__file__).parent.parent / "data" / "trivia.sqlite"
-        questions_path = (
-            Path(__file__).parent.parent.parent / "OpenTriviaQA" / "categories"
-        )
-        load_database(self.db_file, questions_path)
-
-    def invoke(self, msg: str, node: str):
-        self.run_in_thread(self.play, msg, node)
-
-    def play(self, msg: str, node: str):
-        """
-        'trivia' asks a new question
-        'trivia <answer>' checks your answer
-        """
-
-        msg = msg[len(self.command) :].lower().lstrip().rstrip()
-        response = None
-
-        with self.lock:
-            db = sqlite3.connect()
-            cursor = db.cursor()
-
-            game = TriviaGame(cursor, node)
-            response = game.run(msg)
-
-            cursor.close()
-            db.commit()
-
-        if response:
-            self.send_dm(response, node)
-        else:
-            self.send_dm("Game error.", node)
-
-    # def shutdown(self):
-    #     log.debug("Shutting down trivia database.")
-    #     self.database.close()
